refactor to RoomUtilsService
